1월 31일 모든 controller 인증, 인가 추가

+++ src/main/java/SoftSquared/PeopleOfDelivery/controller/CouponController.java
@@ -11,8 +11,11 @@ import SoftSquared.PeopleOfDelivery.provider.CouponProvider;
 import SoftSquared.PeopleOfDelivery.provider.MenuProvider;
 import SoftSquared.PeopleOfDelivery.service.CouponService;
 import SoftSquared.PeopleOfDelivery.service.MenuService;
+import io.jsonwebtoken.Claims;
 import io.swagger.annotations.ApiOperation;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
@@ -23,6 +26,7 @@ import java.util.stream.Collectors;
 
 import static SoftSquared.PeopleOfDelivery.config.BaseResponseStatus.*;
 
+@Slf4j
 @Controller
 @RequestMapping(value = "/api")
 public class CouponController {
@@ -42,11 +46,26 @@ public class CouponController {
     @ResponseBody
     @RequestMapping(value = "/coupons",method = RequestMethod.GET)
     @ApiOperation(value = "전체 쿠폰 조회 (관리자 기능)", notes = "전체 쿠폰 목록 불러오기")
-    public BaseResponse<List<GetCouponRes>> getCoupons(){
+    public BaseResponse<List<GetCouponRes>> getCoupons(
+            Authentication authentication
+    ){
 
         List<GetCouponRes> getCouponResList;
 
         try{
+
+            if(authentication == null){
+                throw new BaseException(EMPTY_AUTHENTICATION);
+            }
+            Claims claims= (Claims) authentication.getPrincipal();
+            int role = claims.get("role", Integer.class);
+            log.info("전체 쿠폰 조회 " + String.valueOf(role));
+
+            if(role!=100) {
+                throw new BaseException(FAILED_TO_GET_AUTHENTICATION);
+            }
+
+
             getCouponResList = couponProvider.retrieveCouponList();
             return new BaseResponse<>(SUCCESS_READ_COUPON, getCouponResList);
         }catch(BaseException exception){
@@ -60,14 +79,24 @@ public class CouponController {
 
      */
     @ResponseBody
-    @RequestMapping(value = "/coupons/{userId}",method = RequestMethod.GET)
-    @ApiOperation(value = "내 쿠폰 조회" , notes = "내 쿠폰 목록 불러오기")
+    @RequestMapping(value = "/coupons/me",method = RequestMethod.GET)
+    @ApiOperation(value = "내 쿠폰 조회 (회원)" , notes = "내 쿠폰 목록 불러오기")
     public BaseResponse<GetCouponRes> getCoupon(
-            @PathVariable Long userId){
+            Authentication authentication
+            ){
 
         GetCouponRes getCouponRes;
 
         try{
+
+            if(authentication == null){
+                throw new BaseException(EMPTY_AUTHENTICATION);
+            }
+
+            Claims claims= (Claims) authentication.getPrincipal();
+            long userId = claims.get("userId", Long.class);
+            log.info("내 쿠폰 조회 " + userId);
+
             getCouponRes = couponProvider.retrieveCoupon(userId);
             return new BaseResponse<>(SUCCESS_READ_COUPON, getCouponRes);
         }catch(BaseException exception){
@@ -80,9 +109,10 @@ public class CouponController {
      */
     @ResponseBody
     @RequestMapping(value = "/coupons/{userId}",method = RequestMethod.PATCH)
-    @ApiOperation(value = " 쿠폰 수정하기 ", notes = " 쿠폰 수정하기")
+    @ApiOperation(value = " 쿠폰 수정하기 (관리자 권한)", notes = " 쿠폰 수정하기")
     public BaseResponse<GetCouponRes> updateCoupon(
             @PathVariable Long userId,
+            Authentication authentication,
             //0은 plus,1 은 minus
             @RequestParam(value = "plusOrMinus",defaultValue = "true") boolean operationCheck,
             @RequestParam(value = "coupon1000Count",required = false,defaultValue = "0") Integer coupon1000Count,
@@ -92,6 +122,18 @@ public class CouponController {
         GetCouponRes getCouponRes;
 
         try{
+            if(authentication == null){
+                throw new BaseException(EMPTY_AUTHENTICATION);
+            }
+            Claims claims= (Claims) authentication.getPrincipal();
+            int role = claims.get("role", Integer.class);
+            log.info("쿠폰 수정 하기");
+
+            if(role!=100) {
+                throw new BaseException(FAILED_TO_GET_AUTHENTICATION);
+            }
+
+
             getCouponRes = couponService.updateCoupon(
                     userId,operationCheck, coupon1000Count, coupon3000Count,coupon5000Count
             );
@@ -106,9 +148,10 @@ public class CouponController {
      */
     @ResponseBody
     @RequestMapping(value = "/coupons/{userId}",method = RequestMethod.DELETE)
-    @ApiOperation(value = " 쿠폰 초기화 하기 ", notes = " 쿠폰 초기화")
+    @ApiOperation(value = " 쿠폰 초기화 하기 (관리자 권한) ", notes = " 쿠폰 초기화")
     public BaseResponse<GetCouponRes> deleteCoupon(
             @PathVariable Long userId,
+            Authentication authentication,
             //0은 plus,1 은 minus
             @RequestParam(value = "coupon1000Count",required = false,defaultValue = "true") boolean coupon1000Count,
             @RequestParam(value = "coupon3000Count",required = false,defaultValue = "true") boolean coupon3000Count,
@@ -117,6 +160,19 @@ public class CouponController {
         GetCouponRes getCouponRes;
 
         try{
+
+            if(authentication == null){
+                throw new BaseException(EMPTY_AUTHENTICATION);
+            }
+            Claims claims= (Claims) authentication.getPrincipal();
+            int role = claims.get("role", Integer.class);
+            log.info("쿠폰 초기화 하기");
+
+            if(role!=100) {
+                throw new BaseException(FAILED_TO_GET_AUTHENTICATION);
+            }
+
+
             getCouponRes = couponService.deleteCoupon(
                     userId, coupon1000Count, coupon3000Count,coupon5000Count
             );
