+++ src/main/java/SoftSquared/PeopleOfDelivery/controller/OrderController.java
@@ -8,8 +8,11 @@ import SoftSquared.PeopleOfDelivery.domain.order.GetOrderRes;
 import SoftSquared.PeopleOfDelivery.domain.order.PostOrderRes;
 import SoftSquared.PeopleOfDelivery.provider.OrderProvider;
 import SoftSquared.PeopleOfDelivery.service.OrderService;
+import io.jsonwebtoken.Claims;
 import io.swagger.annotations.ApiOperation;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
@@ -17,6 +20,7 @@ import java.util.List;
 
 import static SoftSquared.PeopleOfDelivery.config.BaseResponseStatus.*;
 
+@Slf4j
 @Controller
 @RequestMapping(value = "/api")
 public class OrderController {
@@ -37,11 +41,26 @@ public class OrderController {
     @ResponseBody
     @RequestMapping(value = "/orders",method = RequestMethod.GET)
     @ApiOperation(value = "전체 주문내역 조회", notes = "전체 주문내역 조회")
-    public BaseResponse<List<GetOrderRes>> getOrders(){
+    public BaseResponse<List<GetOrderRes>> getOrders(
+            Authentication authentication
+    ){
 
         List<GetOrderRes> getOrderResList;
 
         try{
+
+            if(authentication == null){
+                throw new BaseException(EMPTY_AUTHENTICATION);
+            }
+            Claims claims= (Claims) authentication.getPrincipal();
+            int role = claims.get("role", Integer.class);
+
+            log.info("전체 주문 내역 조회");
+
+            if(role != 100) {
+                throw new BaseException(FAILED_TO_GET_AUTHENTICATION);
+            }
+
             getOrderResList = orderProvider.retrieveOrders();
             return new BaseResponse<>(SUCCESS_READ_ORDERLIST, getOrderResList);
         }catch(BaseException exception){
@@ -57,11 +76,25 @@ public class OrderController {
     @RequestMapping(value = "/order-detail/{orderId}",method = RequestMethod.GET)
     @ApiOperation(value = "주문 내역 상세조회", notes = "주문 내역 상세 조회")
     public BaseResponse<GetOrderDetailRes> getOrderDetail(
-            @PathVariable Long orderId){
+            @PathVariable Long orderId,
+            Authentication authentication){
 
         GetOrderDetailRes getOrderDetailResList;
 
         try{
+
+            if(authentication == null){
+                throw new BaseException(EMPTY_AUTHENTICATION);
+            }
+            Claims claims= (Claims) authentication.getPrincipal();
+            int role = claims.get("role", Integer.class);
+
+            log.info("주문 내역 상세조회");
+
+            if(role != 1) {
+                throw new BaseException(FAILED_TO_GET_AUTHENTICATION);
+            }
+
             getOrderDetailResList = orderProvider.retrieveOrderDetail(orderId);
             return new BaseResponse<>(SUCCESS_READ_ORDER_DETAIL, getOrderDetailResList);
         }catch(BaseException exception){
@@ -73,14 +106,28 @@ public class OrderController {
      * 회원 주문 내역 조회
      */
     @ResponseBody
-    @RequestMapping(value = "/orders/{userId}",method = RequestMethod.GET)
+    @RequestMapping(value = "/orders/me",method = RequestMethod.GET)
     @ApiOperation(value = "회원 주문내역 조회", notes = "회원 주문내역 조회")
     public BaseResponse<List<GetOrderRes>> getUserOrders(
-            @PathVariable Long userId){
+            Authentication authentication){
 
         List<GetOrderRes> getOrderResList;
 
         try{
+            if(authentication == null){
+                throw new BaseException(EMPTY_AUTHENTICATION);
+            }
+            Claims claims= (Claims) authentication.getPrincipal();
+            int role = claims.get("role", Integer.class);
+            long userId = claims.get("userId",Integer.class);
+
+            log.info("회원 주문내역 조회");
+
+            if(role != 1) {
+                throw new BaseException(FAILED_TO_GET_AUTHENTICATION);
+            }
+
+
             getOrderResList = orderProvider.retrieveOrderList(userId);
             return new BaseResponse<>(SUCCESS_READ_ORDERLIST_BY_USER, getOrderResList);
         }catch(BaseException exception){
@@ -95,8 +142,9 @@ public class OrderController {
     @RequestMapping(value = "/orders",method = RequestMethod.POST)
     @ApiOperation(value = "주문하기 (회원 기능)", notes = "주문하기")
     public BaseResponse<PostOrderRes> createOrder(
+            Authentication authentication,
             @RequestParam(name = "requestContent") String requestContent,
-            @RequestParam(name = "userId") Long userId,
+            //@RequestParam(name = "userId") Long userId,
             @RequestParam(name = "storeId") Long storeId,
             @RequestParam(name = "address",required = false) String address,
             @RequestParam(name = "orderPrice") Integer orderPrice,
@@ -112,6 +160,20 @@ public class OrderController {
         PostOrderRes postOrderRes;
 
         try{
+
+            if(authentication == null){
+                throw new BaseException(EMPTY_AUTHENTICATION);
+            }
+            Claims claims= (Claims) authentication.getPrincipal();
+            int role = claims.get("role", Integer.class);
+            long userId = claims.get("userId",Integer.class);
+
+            log.info("주문하기");
+
+            if(role != 1) {
+                throw new BaseException(FAILED_TO_GET_AUTHENTICATION);
+            }
+
             postOrderRes = orderService.createOrder(
                     requestContent,userId,address,storeId,orderPrice,deliveryFee,BasketId,pgName,pgType,pgData,couponType);
             return new BaseResponse<>(SUCCESS_READ_ORDER, postOrderRes);
@@ -127,11 +189,25 @@ public class OrderController {
     @RequestMapping(value = "/orders/{orderId}",method = RequestMethod.DELETE)
     @ApiOperation(value = "회원 주문 삭제하기", notes = "회원 주문 삭제하기")
     public BaseResponse<DeleteOrderRes> DeleteOrder(
-            @PathVariable Long orderId){
+            @PathVariable Long orderId,
+            Authentication authentication){
 
         DeleteOrderRes deleteOrderRes;
 
         try{
+
+            if(authentication == null){
+                throw new BaseException(EMPTY_AUTHENTICATION);
+            }
+            Claims claims= (Claims) authentication.getPrincipal();
+            int role = claims.get("role", Integer.class);
+
+            log.info("회원 주문 삭제하기");
+
+            if(role != 1) {
+                throw new BaseException(FAILED_TO_GET_AUTHENTICATION);
+            }
+
             deleteOrderRes = orderService.DeleteOrder(orderId);
             return new BaseResponse<>(SUCCESS_DELETE_ORDER, deleteOrderRes);
         }catch(BaseException exception){
